# Allow caching of compiler version and argument information to improve startup time