--- a/CONTRIBUTORS.md
+++ b/CONTRIBUTORS.md
@@ -73,3 +73,4 @@ From oldest to newest contributor, we would like to thank:
 - [Sebastian Staffa](https://github.com/Staff-d)
 - [Andreas Jonson](https://github.com/andjo403)
 - [Sam Clegg](https://github.com/sbc100)
+- [Austin Morton](https://github.com/apmorton)