--- a/lib/base-compiler.js
+++ b/lib/base-compiler.js
@@ -103,6 +103,22 @@ class BaseCompiler {
         return exec.execute(compiler, args, options);
     }
 
+    getCompilerCacheKey(compiler, args, options) {
+        return {mtime: this.mtime, compiler, args, options};
+    }
+
+    async execCompilerCached(compiler, args, options) {
+        const key = this.getCompilerCacheKey(compiler, args, options);
+        let result = await this.env.compilerCacheGet(key);
+        if (!result) {
+            result = await exec.execute(compiler, args, options);
+            if (result.okToCache)
+                this.env.compilerCachePut(key, result);
+        }
+
+        return result;
+    }
+
     getDefaultExecOptions() {
         return {
             timeoutMs: this.env.ceProps("compileTimeoutMs", 7500),
@@ -877,7 +893,7 @@ Please select another pass or change filters.`;
         const execOptions = this.getDefaultExecOptions();
         const versionFlag = this.compiler.versionFlag || '--version';
         execOptions.timeoutMs = 0; // No timeout for --version. A sort of workaround for slow EFS/NFS on the prod site
-        return this.exec(this.compiler.exe, [versionFlag], execOptions);
+        return this.execCompilerCached(this.compiler.exe, [versionFlag], execOptions);
     }
 
     initialise() {