--- a/lib/compilation-env.js
+++ b/lib/compilation-env.js
@@ -43,6 +43,8 @@ class CompilationEnvironment {
         this.cache = FromConfig.create(doCache === undefined || doCache ? this.ceProps('cacheConfig', '') : "");
         this.executableCache = FromConfig.create(doCache === undefined ||
             doCache ? this.ceProps('executableCacheConfig', '') : "");
+        this.compilerCache = FromConfig.create(doCache === undefined ||
+            doCache ? this.ceProps('compilerCacheConfig', '') : "");
         this.compileQueue = new Queue(this.ceProps("maxConcurrentCompiles", 1), Infinity);
         this.reportCacheEvery = this.ceProps("cacheReportEvery", 100);
         this.multiarch = null;
@@ -87,6 +89,20 @@ class CompilationEnvironment {
     cachePut(object, result, creator) {
         const key = BaseCache.hash(object);
         return this.cache.put(key, JSON.stringify(result), creator);
+    }
+
+    async compilerCacheGet(object) {
+        const key = BaseCache.hash(object);
+        const result = await this.compilerCache.get(key);
+        if (!result.hit) {
+            return null;
+        }
+        return JSON.parse(result.data);
+    }
+
+    compilerCachePut(object, result, creator) {
+        const key = BaseCache.hash(object);
+        return this.compilerCache.put(key, JSON.stringify(result), creator);
     }
 
     executableGet(object, destinationFolder) {