--- a/lib/compilers/argument-parsers.js
+++ b/lib/compilers/argument-parsers.js
@@ -66,7 +66,7 @@ class BaseParser {
     }
 
     static getOptions(compiler, helpArg) {
-        return compiler.exec(compiler.compiler.exe, [helpArg]).then(result => {
+        return compiler.execCompilerCached(compiler.compiler.exe, [helpArg]).then(result => {
             let options = {};
             if (result.code === 0) {
                 const optionFinder = /^\s*(--?[a-z0-9=+,[\]<>|-]*)\s*(.*)/i;
@@ -153,7 +153,7 @@ class ISPCParser extends BaseParser {
     }
 
     static getOptions(compiler, helpArg) {
-        return compiler.exec(compiler.compiler.exe, [helpArg]).then(result => {
+        return compiler.execCompilerCached(compiler.compiler.exe, [helpArg]).then(result => {
             let options = {};
             if (result.code === 0) {
                 const optionFinder = /^\s*\[(--?[a-z0-9=+,{}\s[\]<>|-]*)\]\s*(.*)/i;
@@ -231,7 +231,7 @@ class VCParser extends BaseParser {
     }
 
     static getOptions(compiler, helpArg) {
-        return compiler.exec(compiler.compiler.exe, [helpArg]).then(result => {
+        return compiler.execCompilerCached(compiler.compiler.exe, [helpArg]).then(result => {
             let options = {};
             if (result.code === 0) {
                 const optionFinder = /^\s*(\/[a-z0-9=:+#.,[\]{}<>|_-]*)\s*(.*)/i;