--- a/lib/handlers/compile.js
+++ b/lib/handlers/compile.js
@@ -109,12 +109,9 @@ class CompileHandler {
                         logger.debug(`${compiler.id} is unchanged`);
                         return cached;
                     }
-                    return new this.factories[type](compiler, this.compilerEnv)
-                        .initialise()
-                        .then(compiler => {
-                            if (compiler) compiler.mtime = res.mtime;
-                            return compiler;
-                        });
+                    const compilerObj = new this.factories[type](compiler, this.compilerEnv);
+                    compilerObj.mtime = res.mtime;
+                    return compilerObj.initialise();
                 })
                 .catch(err => {
                     logger.warn(`Unable to stat ${compiler.id} compiler binary: `, err);