add something like the following to `compiler-explorer.local.properties`
```
compilerCacheConfig=OnDisk(./lib/storage/data/cache, 50)
```

Startup when the cache is empty is marginally slower, but subsequent startups can be an order of magnitude faster, because no compilers will be executed during startup.

With 70 compilers installed locally the reported startup time went from 11604ms to 1711ms.
With 3 compilers installed locally the reported startup time went from 1742ms to 1429ms.

Definitely useful during development, although in production the usefulness really depends on how fast S3 caching would be.  It may simply be faster to execute the compilers off EFS than to make several hundred round trips to S3.