@mattgodbolt 2019-05-29
This looks great: thanks. I'd bet the round-trip to S3 is still faster than running the compiler. We lose the "precaching" of the compilers in RAM, but I bet we run out of free cache RAM before we finish running all the compilers anyway...