@mattgodbolt 2019-05-29
I'v pushed what I hope is a fix to a branch on CE: https://github.com/mattgodbolt/compiler-explorer/compare/apmorton-features/compiler-info-caching?expand=1